pyproject.toml
[project]
name = "qq-bot"
version = "1.0.0"
description = "基于 Nonebot2 的 QQ 机器人，提供短链生成和随机数功能"
authors = [
    {name = "Bot Developer", email = "developer@example.com"}
]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "nonebot2>=2.0.0",
    "nonebot-adapter-onebot>=2.0.0",
    "httpx>=0.24.0",
    "pydantic>=1.10.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=22.0.0",
    "isort>=5.0.0",
    "flake8>=5.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 88
target-version = ['py38']

[tool.isort]
profile = "black"
line_length = 88

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
